+++ src/test/container-templates/templatesCLICommands.test.ts
@@ -76,7 +76,7 @@ describe('tests packageTemplates()', async function () {
 	});
 
     const cwd = process.cwd();
-    const cliHost = await getCLIHost(cwd, loadNativeModule);
+    const cliHost = await getCLIHost(cwd, loadNativeModule, true);
 
 	let args: PackageCommandInput = {
 		targetFolder: '',